+++ lib/src/shared_map_generic.dart
@@ -82,6 +82,13 @@ class SharedMapGeneric<K, V> implements SharedMapSync<K, V> {
   List<V> values() => _entries.values.toList();
 
   @override
+  List<MapEntry<K, V>> entries() => _entries.entries.toList();
+
+  @override
+  List<MapEntry<K, V>> where(bool Function(K key, V value) test) =>
+      _entries.entries.where((e) => test(e.key, e.value)).toList();
+
+  @override
   int length() => _entries.length;
 
   @override
@@ -147,6 +154,14 @@ class SharedMapCacheGeneric<K, V> implements SharedMapCached<K, V> {
   @override
   List<V> values({Duration? timeout, bool refresh = false}) =>
       _sharedMap.values();
+
+  @override
+  List<MapEntry<K, V>> entries({Duration? timeout, bool refresh = false}) =>
+      _sharedMap.entries();
+
+  @override
+  List<MapEntry<K, V>> where(bool Function(K key, V value) test) =>
+      _sharedMap.where(test);
 
   @override
   int length({Duration? timeout, bool refresh = false}) => _sharedMap.length();