+++ test/shared_map_generic_test.dart
@@ -117,11 +117,19 @@ void main() {
 
       expect(await _asFuture(m1).keys(), equals(['a', 'b', 'c']));
       expect(await _asFuture(m1).values(), equals([333, 2001, 3001]));
+      expect((await _asFuture(m1).entries()).map((e) => (e.key, e.value)),
+          equals([('a', 333), ('b', 2001), ('c', 3001)]));
       expect(await _asFuture(m1).length(), equals(3));
 
       expect(await _asFutureOr(m1).keys(), equals(['a', 'b', 'c']));
       expect(await _asFutureOr(m1).values(), equals([333, 2001, 3001]));
+      expect(await _asFutureOr(m1).entries().toRecords(),
+          equals([('a', 333), ('b', 2001), ('c', 3001)]));
+
       expect(await _asFutureOr(m1).length(), equals(3));
+
+      expect(await _asFutureOr(m1).where((k, v) => v == 333).toRecords(),
+          equals([('a', 333)]));
 
       {
         var cached3 = await _asFutureOr(m1).cached();